LXP-comparer/pyproject.toml
[project]
name = "lxp-comparer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-aiplatform>=1.141.0",
    "openpyxl>=3.1.5",
]
